add volume detail page
