+++ lib/main.dart
@@ -20,6 +20,7 @@ class App extends ConsumerWidget {
     final theme = ref.watch(themeProvider);
     return MaterialApp.router(
       title: 'Fluvita',
+      debugShowCheckedModeBanner: false,
       theme: theme.lightTheme,
       darkTheme: theme.darkTheme,
       themeMode: theme.mode,
