+++ lib/riverpod/router.dart
@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:fluvita/models/volume_model.dart';
+import 'package:fluvita/pages/library/volume_detail_page/volume_detail_page.dart';
 import 'package:go_router/go_router.dart';
 import 'package:fluvita/pages/home/home_page.dart';
 import 'package:fluvita/pages/library/libraries_page/libraries_page.dart';
 import 'package:fluvita/pages/library/series_detail_page/series_detail_page.dart';
 import 'package:fluvita/pages/library/series_page/series_page.dart';
-import 'package:fluvita/pages/library/want_to_read/want_to_read_page.dart';
+import 'package:fluvita/pages/library/want_to_read_page/want_to_read_page.dart';
 import 'package:fluvita/pages/reader/reader_page.dart';
 import 'package:fluvita/pages/settings/settings_page.dart';
 import 'package:fluvita/widgets/navigator_container.dart';
@@ -44,9 +46,14 @@ GoRouter router(Ref ref) {
               path: 'all-series',
             ),
             TypedGoRoute<SeriesRoute>(
-              path: ':libraryId/series',
+              path: 'library/:libraryId',
             ),
-            TypedGoRoute<SeriesDetailRoute>(path: ':seriesId'),
+            TypedGoRoute<SeriesDetailRoute>(
+              path: 'series/:seriesId',
+            ),
+            TypedGoRoute<VolumeDetailRoute>(
+              path: 'volume/:volumeId',
+            ),
           ],
         ),
       ],
@@ -137,6 +144,17 @@ class SeriesDetailRoute extends GoRouteData with $SeriesDetailRoute {
   @override
   Widget build(BuildContext context, GoRouterState state) =>
       SeriesDetailPage(seriesId: seriesId);
+}
+
+class VolumeDetailRoute extends GoRouteData with $VolumeDetailRoute {
+  VolumeDetailRoute(this.$extra) : volumeId = $extra.id;
+
+  final int volumeId;
+  final VolumeModel $extra;
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) =>
+      VolumeDetailPage(volume: $extra);
 }
 
 class SettingsRoute extends GoRouteData with $SettingsRoute {
